+++ CHANGELOG.md
@@ -18,6 +18,8 @@ The changelog format is inspired by [keep-a-changelog](https://github.com/olivie
   apprach using only lengths of real-space cell) made available as
   option through Python interface and CLI.
 - Add emulation of VASP Auto and KSPACING modes.
+- Restructure for packaging with setuptools
+- Installation instructions in README
 
 ## [0.2] - 2016-07-28