+++ CHANGELOG.md
@@ -13,14 +13,21 @@ expected to be especially active and follows a simplified Semantic Versioning:
 The changelog format is inspired by [keep-a-changelog](https://github.com/olivierlacan/keep-a-changelog).
 
 ## [Unreleased]
+
+## [1.0] - 2016-10-07
+
+### Changed
 - Correctly calculate reciprocal lattice vectors using whole
   cell. This is the new default behaviour. Former behaviour ("naive"
   apprach using only lengths of real-space cell) made available as
   option through Python interface and CLI.
-- Add emulation of VASP Auto and KSPACING modes.
-- Restructure for packaging with setuptools
+- Restructure repository for packaging with setuptools
 - Installation instructions in README
-- Add range generator for convergence testing
+
+### Added
+- New emulation of VASP Auto and KSPACING modes.
+- Setuptools installation
+- New range generator for convergence testing: **kgrid-series**
 
 ## [0.2] - 2016-07-28
 
@@ -39,5 +46,6 @@ The changelog format is inspired by [keep-a-changelog](https://github.com/olivie
 - Optparse-based interface
 - GPL
 
-[Unreleased]: https://github.com/wmd-group/kgrid/compare/v0.2...HEAD
+[Unreleased]: https://github.com/wmd-group/kgrid/compare/v1.0...HEAD
+[1.0]: https://github.com/wmd-group/kgrid/compare/v0.2...v1.0
 [0.2]: https://github.com/wmd-group/kgrid/compare/v0.1...v0.2